backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
websockets==13.1
crewai==0.67.1
crewai-tools==0.12.1
aiofiles==24.1.0
python-dotenv==1.0.1
